Implemented failure rollback. Tested it with the following example on UI Need to try and emulate IO and/or COM exception failures
A.1.0 depends on B.1.0 (net40) which depends on nothing
A.2.0 depends on B.2.0 (net40) -> Net45OnlyPackage.1.0 -> C.1.0 -> D.1.0

First, install A.1.0 on a net40 console app
Next, preview update A.1.0 to A.2.0. The list will be as follows

Installing: D.1.0, C.1.0, Net45OnlyPackage.1.0, B.2.0, A.2.0
Uninstalling: A.1.0, B.1.0

Perform update. And, it will uninstall A.1.0, B.1.0. And, install D.1.0
and C.1.0. Will fail to install Net45OnlyPackage.1.0 and throw

And, it will rollback the actions as follows
Uninstall Net45OnlyPackage.1.0
Uninstall C.1.0
Uninstall D.1.0
Install A.1.0
Install B.1.0