Fixed https://github.com/NuGet/Home/issues/183
Changes: We use a local bool IsSolutionOpening which will only be set when
OnSolutionOpened is called and it raises the SolutionOpening event first
before ensuring NuGetProjectCache. VSSettings uses SolutionOpening event
instead of SolutionOpened, so, NuGetProjectCache is never ensured
before ISettings is refreshed to use custom repositoryPath. Also,
EnvDTEProject added and renamed will not do anything until IsSolutionOpen
is true. And, IsSolutionOpen will not be true if IsSolutionOpening is not

RepositoryPath is read correctly when Visual studio is restarted.
This means that NuGet.Configuration is fine. However, when a solution
is re-opened or simply, when 1 solution is closed and another is opened
within the same VS instance, the solution open event comes after the
project added event. This causes the issue for the following reason

VSSettings, is refreshed during solution open or close. So, when a
solution is closed, VSSettings is refreshed, and only global config
is used in the hierarchical config. At this point, before solution
open is received, if projectAdded is received, VSNuGetProjectFactory
tries to create a FolderNuGetProject, and says that the repositoryPath
is the default one

After this, Solution open event is received and the custom repositoryPath
comes into picture now. So, all the installs happen at the wrong packages
folder, but, restore works fine