TO enable Sync but parallel package restore on build. Passing down solution directory to most of IPackageRestoreManager methods. This helps ensure that solution information do not queried from a background thread
Brought back the IsCurrentSolutionEnabledForRestore property and made
EnableCurrentSolutionForRestore to no-op. Also, both methods are not
virtual anymore. In VSPackageRestoreManager, all the methods are gone

Added a PackageRestoreFailed event so that package restore failures are
handled in the PackageManagement layer. This helps have most of the logic
in the code common to VS and nuget.exe

In addition, the code is more testable and have added more tests