On MVC project creation, packages like EntityFramework get installed. And, EntityFramework has install.ps1 which access DTE objects and the powershell scripts are executed in the pipeline execution thread
Since IWizard.ProjectFinishedGenerating is called on the main thread,
and JTF.Run is used on that method, and install.ps1 from EF accesses DTE
from a non-UI thread, there is a deadlock

Using a PumpingJTF which keeps the filtered message pump running. This fixes
the problem since marshalling is still allowed. Will find out if it is
possible to execute Powershell scripts on the main thread after this

Testing done after the fix

1) Tested that MVC web application can be created right after opening VS
without having PMC open
2) Tested that EntityFramework can be installed and uninstalled from the UI
and powershell console
3) After doing 1 and 2, on the same instance, successfully created MVC+WebAPI+WebForms
application
4) Also, tested that Enable-Migrations is available from the powershell console