Fix the NullReferenceException problem on the UI. The problem was caused by a recent change: NuGet/NuGet.PackageManagement@54d3e1d
With this change, right after user clicks the action button, a refresh is triggered. Which in turn will cause a UI refresh. The problem is that during the UI refresh, the DetailControl.ContextData is null, and at this time, accessing properties such as NuGetUI.FileConflictAction will cause NullReferenceException. Note that this is timing related so the problem doesn't happen all the time.

The fix is to read the values on the UI and save them in NuGetUI before starting the package restore.

Fix NuGet/Home#701