Fix NuGet/Home#666
The cause of the problem NuGet uses PackageManagerModel.Context.Projects list, so it doesn't distinguish between
a solution containing just one project, and one project. Fix this by using PackageManagerModel.SolutionName property, which
is non empty when user opens package manager on the solution, while it's null when opening package manager on a project.

Also fixes the label, so that on the solution package manager, the title is: Package Manager: Solution 'solution_name'.