Fix NuGet/Home#668 - Detail pane does not refresh correctly after a package is selected.
The cause of the problem is that the background tasks to load version info in V2 get cancelled.
Without the versions list, only the Id and Icon get refreshed when a package is selected.

The fix is to not use cancellation token in those background tasks. Since we used Lazy<>, those
tasks will only start when the versions data are needed, so this change will not generate lots of
unneccessary tasks.