Fix NuGet/Home#1300: V3 install/restore only uses credentials for index.json.
The problem is that in HttpHandlerResourceV3Provider, when there is a credential saved in
nuget.config, the code createsCredentialCache, then adds NetworkCredential into the cache, with the
uri of the source as key. However, forV3 source, other resources most likely have different URIs
from index.json endpoint. Thus, this CredentialCache won't be able to provide credentials for other
resources. It only works for index.json.

The fix is to just create a normal NetworkCredential object.