Fix NuGet/Home#1336: Msbuild selection in nuget.exe 3.2 RC is broken
nuget runs "msbuild.exe /version" to get the version of msbuild in PATH, then matches it with the
installed versions by major and minor version number. However, the version number of a recent update
of msbuild 4.0 is now 4.6.x.x, which does not match an installed version.

The fix is to do match by major number if matching by major + minor fails. nuget also displays the
selected msbuild version on auto detection.