Fix https://github.com/NuGet/Home/issues/1397
This is probably a first step in a series of possible improvements to uninstall/update.
This change adds a batch operation around removing files, and modifies the batch operation to be more friendly to not require finding out what files change ahead of time

On my laptop I measured the improvement on KendoUICore uninstall from 350 seconds to 120 seconds.