Fixes https://github.com/NuGet/Home/issues/1021.
The issue is that lib folder with just the _._ file is not considered as
a valid 'any' framework. That is, lib/_._ is being ignored. However,
lib/net45/_._ or lib/net45 will be considered as valid.

Added a property to FrameworkSpecificGroup which says whether there was
an empty folder represented by file '_._'.

Even after this change, an empty folder of lib without the '_._' file
is not a valid empty directory and will be ignored.