Fixes https://github.com/NuGet/Home/issues/917
Execute the prompt function from a worker thread, so that the UI thread
is not blocked waiting on it. Note that a default prompt function
as defined in Profile.ps1 will simply return a string "PM>".
This will always work. However, a custom "prompt" function might call
Write-Host and NuGet will explicity switch to the main thread using JTF.
If the main thread was blocked then, it will consistently result in a hang.