Fix https://github.com/NuGet/Home/issues/1412
A server (or package) might incorrectly return multiple dependencies in the same package.
Although this can be ignored using distinct. It indicates that something is very wrong with the package.

We know of at least one server that combines dependencies, which leads to unrelated dependencies and versions in the same dependency group.

Hence the fix is to throw a better error message.