Fixes https://github.com/NuGet/Home/issues/1374
Disabled Machine wide package sources specific to VS, get enabled when
running nuget.exe sources add.

Reason: Machine wide package sources specific to Visual Studio
do not get loaded in the context of nuget.exe. However, they may be
disabled on the NuGet.Config at %APPDATA%. When saving package sources,
only sources that are loaded in the context of the client, are eligible for
being disable. This is incorrect

Fix: Every disabled package source element in a nuget.config
should be added back whether or not it is a loaded package source
in the context of the current client

Added a new unit test for this scenario. Tested it manually. And, tested
that all the existing tests run successfully.