Functional tests revamp
Added a bunch of powershell scripts to do the following and more.
Parameterized the scripts as much as possible
such that nothing is hard-coded.

1) Pick up bits from the given branch, build number and so on
2) Install/Uninstall VSIX
3) Launch VS and start the functional tests

Implemented a bootstrap.ps1 that will used by CI to
copy and extract EndToEnd.zip, which contains all the necessary scripts to
run the functional tests

Added a copy tools and setup machine script to copy the necessary CI tools
and to set up registry keys specific to the CI machine. This is different
from the setup machine script intended to be run on the dev boxes

Install NuGet VSIX script does the following

1) Kills any running instances of VS
2) Uninstalls NuGet VSIX if installed
3) Installs the NuGet VSIX as indicated by $NuGetDropPath after copying
it to $FuncTestRoot

Launches VS and waits until DTE has become available. Controlled by
polling timeout and frequency

Runs the functional tests after obtaining DTE which is guaranteed
to be available. Also, provides real time monitoring. Finally, copies the
test results to the CI