Allow overriding CommonApplicationData on Linux
Red Hat has requested a way to override the location of nuget
configuation data due to how they plan to package components on their
end. This allows them to set an environment varible as the root of the
common application data folder, which will also control where the
machine wide settings come from.