Fix duplicate dependencies in project.lock.json
- Today when we encounter a framework assembly with the same name as a
package we prefer the reference assembly over the packages. Today we do
this incorrectly and as a result, the resolved graph ends up with
multiple versions of packages in the project.lock.json. During conflict
resolution consider the "best" version to be the reference assembly
which will mark all other nodes as rejected.
- Added tests