Issue 2335: NuGet.exe not FIPS compliant.
Currently, SHA512.Create() will always use the SHA512Managed class, which is not FIPS compliant and will throw an exception when FIPS compliance is turned on in Windows.

We have code in Configuration.CryptoHashProvider that uses the appropriate class when FIPS compliance is turned on. I've moved that code to Common, and switched Packaging.PackageExtractor to use it.

This resolves https://github.com/NuGet/Home/issues/2335