Content files in referenced projects are not always included
If there are no content files in the main project but there are content file in a referenced project, they will be missed during pack.  Also, if there are content files in a project file without referenced projects but not in the nuspec, the content files will be missed.  This change takes out the nuspec check and just always looks for content files so that they aren't ever missed.