Fixing nuspec pack with empty Files element so it doesn't include other files
The logic to include other files was Files.Any() in 3.4 but should check for whether the Files element even existed.  This change creates a bool HasFilesNode that tracks whether or not there was a Files node.  This keeps the code from having to check Files for null in many places.