Removing auto fallback to v1 lock files
This change removes automatic fallback of v2 to v1 lock files based on the types of projects used. The previous check relied on xproj to bump the version to v2, however with xproj going away this can no longer be used. The fallback was originally added as a temporary workaround for users with the latest version of NuGet who were still on Visual Studio 2015 Update 1. With the release of Updates 2 and 3 this is no longer needed.

Fixes https://github.com/NuGet/Home/issues/3011