Avoid duplicate nuspec files during v3 extraction
The V3 folder format requires nuspec files to be named with the package id. This change fixes an issue where the nuspec is copied but the original nuspec is also extracted. This causes problems later when using the package folder reader to read the nuspec.

This fix also adds GetNuspecFile to the package reader instead of an extension method. The previous check did not filter out non-root nuspec files and could have found extra duplicates that should have been ignored. It also improves performance for the package folder reader which has an optimized path for finding the nuspec on disk.

Fixes https://github.com/NuGet/Home/issues/3231