allow nuspec files element to override what content files are added while packing using csproj (#860)
Fixes: NuGet/Home#3257

This bug was introduced because of fixing the bug from NuGet/Home#2658 commit 10a986f

manifest.Files != null was not a good enough check as Files was always initialized to an empty list upon creation of a Manifest object. This replaces that check with the better manifest.HasFilesNode which knows if the element is present or not in the nuspec. If an empty files element is present in the nuspec, then content is not added at all .

Since this is a regression, this fix will also go into 3.5-rtm branch.