Properly request DTE service from the components service provider
The previous approach failed if the MEF component happened to be
instantiated from a background thread (i.e. another MEF component
requested in a background thread depended these exports). Since
MEF components aren't guaranteed to initialize on the UI thread,
this caused heard to track and unfriendly errors, such as
NuGet/Home#3419.

The requirement came from the helper class which used a static
service locator and therefore had no way of knowing if it was
being properly invoked, and therefore resorted to just throwing
if not.

This fix changes the way these services are requested from the
MEF components so that they use the built-in mechanism provided
by VSMEF to retrieve VS services properly from a MEF component:
the IServiceProvider import.

Since the helper does not need to retrieve the service itself
anymore, it receives the DTE as a parameter instead.

This fixes NuGet/Home#3419

The previous issue can be easily reproduced with the following
VS package code:

	[ProvideAutoLoad(UIContextGuids.SolutionExists)]
	[PackageRegistration(UseManagedResourcesOnly = true)]
	[Guid(PackageGuidString)]
	public sealed class NuGetBreakingPackage : Package
	{
		public const string PackageGuidString = "3FF684D3-61BF-417C-A1C0-17CD9B81BFE6";

		protected override void Initialize()
		{
			base.Initialize();

			var components = (IComponentModel)GetService(typeof(SComponentModel));

			System.Threading.Tasks.Task.Run(() =>
			{
				try
				{
					var solution = components.DefaultExportProvider.GetExportedValue<object>("NuGet.PackageManagement.ISolutionManager");
				}
				catch (Exception ex)
				{
					Debug.Fail(ex.ToString());
				}
			});
		}
	}