Create FallbackFramework for PackageSpecs when reading msbuild output
This fix add imports to the project target framework for the in memory instace of a PackageSpec. Previously there were no scenarios where the PackageSpec was created and then used in memory, it was always written and read from disk. Loading the file from disk added the imports.

Fixes https://github.com/NuGet/Home/issues/3621