MSBuild project reference support
This change adds a frameworks section under restore metadata, project references are now placed there instead of in both the restore section and the dependencies section. This allows projects to be referenced with a unique name or path, and later resolved to the correct PackageId or project name. The project name can only be known after reading all projects.

An additional benefit of splitting up the references between what was provided by msbuild and what is in the spec is that UWP can continue to work exactly as it has in the past, while .NET Core projects can take advantage of the restore metadata.

Handling for missing projects is also improved with this change. Resolving a project path to the actual project is now handled in the resolver which allows it to fail in the same way as a missing package.

Fixes https://github.com/NuGet/Home/issues/3611
Related to https://github.com/dotnet/sdk/issues/200