Write restore outputs to temp files and then move them
This change writes the assets files, targets, and props to a temporary file location first. Then moves the file into the intended location. This is done to reduce the risk of an external process trying to read a file as it is being written out by restore. It also reduces the chance of multiple nuget restores conflicting with each other.

Fixes https://github.com/NuGet/Home/issues/3806