Project compile/runtime asset support
This change adds compile/runtime assets to project entries in the assets file. Transitive projects will keep their assets and non-transitive projects will be filtered to _._

Added a generic files property to project entries, this can be used as the base for eventually adding all assets that would be packed into a nupkg for a project into the restore entry. A file entry contains both the package path and the file path on disk.

During restore project assets are read and filtered in the same way package assets are consumed.

The current files used for compile are place holders and do not represent the actual project output. This is expected on the build task side and matches the previous xproj behavior.

PackageSpecs from Non-NuGet projects will now flow into the resolver as part of the files change. Support is still needed for getting the correct TFM for these projects, but we are now very close the resolver being able to work with all projects.

Fixed support for Include/Exclude on Project references.
Fixed recursive restore support in nuget.exe

Fixes https://github.com/NuGet/Home/issues/4076