Improve failed project reference item
This message can occur when any item in a project fails to resolve and Visual Studio DTE is unable to return it. This is a problem if the item is a project reference, but the rest of the time it does not impact NuGet. Based on the current feedback this message is appearing much more frequently and leading users to believe that there was a restore error when there was not.

The intent of the actual message is just to inform users that if they do see missing packages that it could have been caused by a broken project reference.

The fix for this is to move the message from warning to verbose. If a user does hit the rare scenario of getting an incomplete restore they can debug it by increasing their verbosity and reading the new, more helpful, message.

Fixes https://github.com/NuGet/Home/issues/3955