Do NOT open output window during auto-restore
Resolves NuGet/Home#4274.

Current behavior for auto-restore was to activate output console on
first message, e.g. on header line "Restoring NuGet packages...".
In order to stop annoying the user with output window the `Activate`
method call was removed from this scenario. The ouput window itself
will be created and log messages will be logged after this change. In
order to see them user will have to locate the output pane titled
"Package Manager" and make it visible.

Other scenarios, including on-build and explicit solution restore, will
NOT be affected.