Avoid using EntityExists LSL API
Resolves NuGet/Home#5391.

`IndexService.EntityExists` API is not reliable as indexing data may not
be available for a long period of time during cold start of LSL.

This change streamlines provider conditionals for PackageRef project
creation in both deferred and non-deferred modes. In summary, PackageRef
project is to be created when underlying project explicitly defines
`RestoreProjectStyle` and/or contains package reference(s).