Change NU1701 to "This package may not be fully compatible with your project."
The previous message "This may cause compatibility problems" felt ominous. As if I had done something wrong.

The new message is designed to indicate that the package could be partially, or fully, compatible. Also it re-iterates that the package is the potential source of incompatibility. This to mean comes across as saying "this isn't necessarily a problem"

This resolves NuGet/Home#5538