Resolve VS hang when adding a UWP project to a solution with WPF project
Resolves internal bug 470522.

This PR changes Lazy mode to avoid a deadlock in `NuGetUIThreadHelper` caused by CPS trying to create `ProjectServiceHost`.
`PublicationOnly` mode lets parallel threads execute value factory method without being blocked on each other.
It is correct behavior in this case as the value factory provides the same value each time it is called and Lazy is used just for caching the value for perf reasons.