Additional circular dependency detection
This adds an additional check during the walk to mark cicular dependencies. In the scenario that a project references a package with the same name the current predicate chain check is not evaluated, this extra catches those scenarios and allows restore to fail with a helpful message.

Fixes https://github.com/NuGet/Home/issues/5630