Fixing install command test failure on mono (#1655)
This test was added as part of https://github.com/NuGet/NuGet.Client/commit/6f8d529312187fa1b6734da8d4ed9dc3442b916c

The test's logic was correct, but the paths used to assert were not right for mono. I have fixed that and cleaned the test a bit.