Exclude frameworks with invalid assets from compat suggestions
A package fails the compat check displays a message containing a list of compatible frameworks as a suggestion to the user. Groups that do not contain any valid items should be excluded from this set since they do not count as compatible items.

Fixes https://github.com/NuGet/Home/issues/4259