Fixing VsSolutionRestoreService.ToPackageSpec to have same project name in PackageSpec and RestoreMetadata (#1754)
## Bug
Link: https://github.com/NuGet/Home/issues/6020
Regression: Yes

## Fix
Details: In 15.4 we introduced transitive no warn which used the Package spec and the restore metadata to traverse the dependency graph. This exposed a bug in `VsSolutionRestoreService.ToPackageSpec` where, if the project had an `AssemblyName` different from the project file name, the `PackageSpec.Name` and `PackageSpec.RestoreMetadata.ProjectName` were being assigned different values. This caused transitive no warn to throw a key not found exception.

## Testing/Validation
Tests Added: Yes
Validation done:  manual validation done on a project that was failing.