Support NuGet PackageReference properly in temporary assembly by enabling .g.targets and .g.props to be properly included, despite the project file name change.
We're copying GenerateTemporaryTargetAssembly class from PresentationBuildTasks.dll and creating a GenerateTemporaryTargetAssembly2 class.
The temp proj file will no longer be:
<randfilename>.tmp_proj
it will now be:
<randfilename>.<originalProjectExtension>

And  we'll set _OriginalProjectName msbuild property so that microsoft.common.props/targets can find the <origProjectName>.g.props/targets.

There are a set of changes necessary with msbuild as well.