Adding support for Oreo style signing (#1818)
## Bug
Fixes: https://github.com/NuGet/Home/issues/6175

## Fix
Details: This PR adds support for signing format which includes adding signature file to the package. The verification changes to calculating the hash of a signed package excluding the signature file.  
Things to Note - 
1. This removes ZipSigningUtility references through out the solution.
2. Moving all the packaging tests to NuGet.Packaging.FuncTests. So some of the changes are related to moving files etc.

Finally, this PR does not address the following which should be done separately - 

1. Resolve some open questions about signed package archive - https://github.com/NuGet/Home/issues/6211
2. Add support for SignedPackageArchive verification if the signature is not the last entry in the zip -
 https://github.com/NuGet/Home/issues/6212


## Testing/Validation
Tests Added: Yes
Validation done:  Manual validation using sign and verify commands