Improve Package Spec cloning  (#1796)
The Package Spec is cloned 2 times per project during one no-op restore. In VS that's done 3x.
Each call takes 8-9ms.
This change cuts it to ~1ms per call.

As mentioned in the comments earlier, I did not want to use existing APIs, because they did not imply deep cloning, while it's very important that the package spec is deep cloned.

Some types like NuGetFramework and similar are immutable so those are/will not be cloned.