Switching NuGet.Build.Tasks.Pack packing to nuspec (#1888)
As part of https://github.com/NuGet/NuGet.Client/pull/1772 and https://github.com/NuGet/NuGet.Client/pull/1778 @nkolev92  switched the packing of `NuGet.Build.Tasks.Pack` to using the csproj file instead of a nuspec file.

This creates an issue that surfaced while working on https://github.com/dotnet/sdk/pull/1829. Because of using csproj file the generated package gets package dependencies, which is not accurate as the pack dll is ILmerged.

This change reverts the packing of the project to using a nuspec file. 
**This will need to be cherry-picked onto the `release-4.6.0-preview2` branch for an updated insertion into cli/sdk**. 
VS insertion is fine as it does not consume this package.