Fixed an intermittently failing e2e test
Sometimes this test InstallPackageWithNonExistentFrameworkReferences fails because underline error message from Project System is different which doesn't change anything for NuGet. For us, it just failed to add reference which is what we should be checking for. So added a new PS api Assert-Contains which will match if actual exception message has matching text for "failed to add reference..."