Remove VisualStudioVersion check from IsNonInteractive for nuget.exe
NuGet.exe will now prompt for credentials when VisualStudioVersion is set. This was added as a workaround for nuget.exe restores that ran from VS in NuGet 2.x when prompting was first added. At this point it no longer makes sense to block this, and it leads to confusing behavior for users running in the developer command prompt.

Fixes https://github.com/NuGet/Home/issues/6385