Log errors from PackageReference project operations to the error list
In addition to the generic restore failed rolling back message all errors from restore should be displayed in the error list. Errors/warnings are also displayed to the restore output window, this is to make it more clear what the root cause of the failure was.

Fixes https://github.com/NuGet/Home/issues/6427