Consume new `References3.AddFiles` API
This change adds basic usage of the new `References3.AddFiles` API when
adding references in a packages.config-based project. The current API
used for this purposes, `References.Add`, can trigger multiple
design-time builds per call and may not set all the right metadata on
the resulting `<Reference>` element in the project file. This can
require further calls to update that metadata, which can in turn trigger
further design-time builds.

`AddFiles` on the other hand creates `<Reference>` items with all the
proper metadata--notable, the `<Private>` metadata is set appropriately,
and it specifies an explicit version of the given assembly. And since it
is explicitly about adding files as references and does not need to
support short or full assembly names it can skip several of the
design-time builds run by `Add`. Use of `AddFiles` instead of `Add`
eliminates up to 75% of the design-time builds incurred while adding a
new package.

For the time being, if `AddFiles` fails we will fall back to calling
`Add`. This way adding a new package is not completely broken in case
there are bugs in the `AddFiles` implementation. If the new API proves
reliable, then the fallback can be removed later.

`AddFiles` can also take multiple file paths, and resolve all the
assemblies with a single design-time build. However, using the API in
that manner will require some refactoring of the client code; it is an
option for the future but is not pursued here.