Telemetry for invalid casts
Report a simple telemetry event if the cast to `References3` fails. This
indicates that we stumbled upon a project system that doesn't implement
this interface yet, or we have a setup authoring where the
VSLangProj157.dll primary interop assembly isn't installed properly.
This way we can find and correct the issues while still handling the
exception and falling back to using the `Reference` interface.