Fixing string comparison in ProjectRestoreMetadata.Equals (#2166)
As part of a [commit](https://github.com/NuGet/NuGet.Client/commit/c318e14bac14895497080990baacbb777a33bf8b) recently we changed the assets file comparison to use OS based comparer. It seems we forgot to pass the comparer  to `EqualityUtility.OrderedEquals` for sequence comparison. This would make it default to `StringComparer.Ordinal`.

This PR fixes that and adds tests for equality comparison and `EqualityUtility`.