Add env var to filter auth types for configured creds
Currently, it is very difficult to interact with servers that support
Negotiate authentication if default credentials are not the right
credentials. With dotnet on Linux, it is currently impossible. On
Windows, Negotiate itself can work, but if the logged-in Windows user
isn't the right user, results can be counterintuitive, and if you
actually do need to present credential via the Basic mechanism,
you're out of luck unless you can write a credential provider (and that
option isn't even available with .net core nuget)

I think this is a bug in the library, but with .net core on Linux, .net never falls
back to Basic if it has a credential for Negotiate, even though it also
doesn't try to actually perform Negotiate auth, even if an explicit
credential is given.

This change extends the mechanism used to filter auth types for
credential providers to work with configured credentials as well. The
user will specify NUGET_AUTHENTICATION_TYPES=basic
(NUGET_CREDENTIAL_PROVIDER_OVERRIDE_DEFAULT=true may also be needed for
.net framework-based NuGet to be offered the chance of not using default
credentials). This will cause NuGet to respond "don't have any" when
asked for credentials for Negotiate, NTLM, etc, forcing .net to fall
back to Basic.