Ensure Package ID is logged into assets file for signature errors/warnings (#2295)
Currently package signature verification logs do not indicate what package caused the failure. This PR improves this by adding package identity to the assets file in case of a restore failure.

In the PR `SignatureLog` has been updated to be of type `ILogMessage` and a helper method has been added to transform a `SignatureLog` into a `RestoreLogMessage`, which is then written into the assets file.