Extending PumpingJTF from DelegatingJoinableTaskFactory instead of JoinableTaskFactory to resolve CPS deadlock (#2351)
Unfortunately, what CPS needs is the callback to OnTransitioningToMainThread which is on the JoinableTaskFactory (and a switch to main thread calls that only for the ambient jobs owner, which in this case is the PumpingJTF and not the CPS JTF). So using the context is not enough, the full JTF must be used. The solution is to wrap and delegate to an inner JTF, which is why we've moved to DelegatingJoinableTaskFactory which allows passing a full JTF instead of just context.