Don't silently ignore exception from AssemblyMetadataExtractor.ExtractMetadata (#2369)
I just wasted many hours trying to figure out why replacement tokens were not working with my .csproj/.nuspec.

The comments in the source code explain the problem. I'm using UWP, which uses .NET core, so it appears that it is not expected that it should work in my case. However, this doesn't appear to be documented anywhere other than this one comment in the source code.

There appears to be a related issue where a malformed version string in AssemblyInformationalVersion also causes an exception which is also silently ignored in this catch statememt.

Let's print a message here, at least when verbose logging is enabled so that users can at least get some hint of what is going on without having to dig into the source code.

Related issues: #5548, #5807, #4234 (and a bunch of closed ones too).