Removing NU5113 since it can never be invoked (#2385)
This PR removes `NU5113` since it cannot happen. The warning was thrown if a generated package was not SemVer compatible. However, since this is only called right after package generation, it will never run into the case where the package version is not SemVer compatible since we always create SemVer compatible versions.