Fix deadlock in NuGet SDK resolver
The legacy project system in Visual Studio evaluates projects on the UI thread.  When an SDK NuGet package cannot be located, the threading ends up deadlocking and hanging Visual Studio.

By running the async task in a Task.Run() it executes on a different thread and accessing .Result will not deadlock.

https://devdiv.visualstudio.com/DevDiv/_workitems/edit/664977?src=WorkItemMention&amp%3Bsrc-action=artifact_link