Improve perf when searching in a local package source.
When searching for a package in a large Local/UNC package source, the LocalPackageSearchResource would enumerate the entire feed and parse all folders into LocalPackageInfo, and only then filter out anything that doesn't match the search terms.
This change greatly improves the performance (minutes -> sub-second on my machine with my large UNC package feed) of the search by avoiding parsing packages which don't match the search terms.
It is possible to improve perf further by parallelizing the searches for every term, but sub-second is good enough for my purposes.