Initialize ILMerge with appropriate reference assemblies
Previously, ILMerge was always default-initialized with current machine's
C:\Windows\Microsoft.NET\Framework\v4.0.30319. Furthermore, in the .NET
Standard case, it was mixed with implementation assemblies on /lib from
a one-off CoreCLR implementation. It was pure luck that the resulting
assembly used to work.

This also removes the one-off CLI that carries the one-off CoreCLR from
the build altogether.