Move cross-plat libraries to .NET Standard 2.0 (#2118)
* Move xplat libraries to netstandard2.0 and netcoreapp2.1

* Update dotnet cli to 2.2.200

* Remove CandidateAssemnblyFiles from RAR search paths
Microsoft.Build.Runtime package has an mscorlib.dll content file that
ends up winning over mscorlib.dll in .NET Framework targeting pack
mscorlib.dll if CandidateAssemblyFiles are searched.

Legacy project search paths are left unchanaged because they were not
impacted by this and some of the rely on resolution in GAC or
AssemblyFolders.

* Initialize ILMerge with appropriate reference assemblies

Previously, ILMerge was always default-initialized with current machine's
C:\Windows\Microsoft.NET\Framework\v4.0.30319. Furthermore, in the .NET
Standard case, it was mixed with implementation assemblies on /lib from
a one-off CoreCLR implementation. It was pure luck that the resulting
assembly used to work.

This also removes the one-off CLI that carries the one-off CoreCLR from
the build altogether.

* Disabled failing signing tests on macOS

* Remove explicity reference to system.collections.immutable

* Add the ability to insert to multiple CLI/SDK branches