Add flag to skip MSBuild step during restore
The MSBuild analysis of the projects is expensive on large projects (17 seconds on our 400-project solution). For legacy solutions (like ours) where all package references are handled in the old-style packages.config files, the MSBuild analysis is unnecessary and just wastes time on every package restore, even those that should be no-ops. Allow callers to pass a flag that skips this step.