Support custom HttpHandlerResource provider
Running a NuGet restore or install with a project that uses
PackageReferences would bypass the custom HttpHandlerResources
associated with the SourceRepository instances. Whilst most operations
use the HttpHandlerResource provider some operations that use the
CachingSourceProvider result in new providers being created.

Now there is a way to specify a custom provider for the
HttpHandlerResource.

Fixes: https://github.com/NuGet/Home/issues/8330