Refactor logic in _GetRestoreProjectStyle target to a task (part 2)
This is a reworking of my original change but this attempt should not break the Mono scenario.

This change only touches the task assembly and not any common assembly.  In the Mono scenario, an older version of NuGet.Commands.dll is loaded and when the newer logic would call into it we'd get MissingMethodExceptions.

This assembly is not loaded by anything but MSBuild so we know its always going to be the correct version.