Improve degenerative glob filtering case and fix on Linux Nuget wants to remove all .nuspec files from the pack. To do this, it currently adds a filtering pattern of **\*.nuspec. This is then translated into Regex later on. The regex generated for this is pretty complicated and very very slow. Furthermore, the globbing pattern is specific to Windows (it doesn't use Path.DirectorySeparatorChar), and so does not filter on Linux.
Eliminate the \, which still matches the same patterns based on globbing rules and drastically
improves performance. Add a few additional tests too.

This was discovered in an investigation of some packages that were taking a long time to pack in
the dotnet/sdk repo. These packages have nuspecs with a large number of files, and the files
also have long-ish file paths. On my local machine, here are the pack times:

| NuGet Version | Time (sec) (n = 10) | Diff from 5.4.0 |
| ------------- | ------------- | --------------- |
| 5.4.0         | 121.25        | N/A             |
| 5.5.1         | 50.86         | -56.83%         |
| w/ Fix        | 1.48          | -99.98%         |