Move WaitAsync outside try/catch
In case `WaitAsync` throws, `Release` should not be called. Although, `WaitAsync` only is documented to throw `ObjectDisposedException`, which would cause `Release` to throw `ObjectDisposeException` later on, this would make the code resilient to changes to other overloads, which can throw other exceptions too, like `OperationCanceledException` if a cancellation token is passed.