Populate dependencies node subtrees for packages and projects
Populate descendant tree items for top-level package and projects displayed in the "Dependencies" node of projects based on the .NET Project System (which in turn is based on CPS).

Top-level dependencies are populated via project evaluation and design-time build data. Nodes lower in the tree are lazily constructed only if and when the user expands the tree enough for them to be visible, or if the user initiates search in Solution Explorer.

Data is obtained from the `project.assets.json` file, and updated whenever it changes. This data is captured in an immutable `AssetsFileDependenciesSnapshot` object and published via Dataflow in response to other project system data sources.

Tree items are constructed via the following mechanisms:

- Each type of node in the tree has its own 'item' (eg. `ProjectReferenceItem`, `PackageReferenceItem`).
- Children of top-level dependencies are reported via `PackageReferenceAttachedCollectionSourceProvider` and `ProjectReferenceAttachedCollectionSourceProvider`, which are called for all `IVsHierarchy` items in the tree.
- Further children are modelled declaratively via a set of 'relations' (eg. `ProjectToPackageRelation`, `ProjectToProjectRelation`, `PackageToPackageRelation`).

For more information about the APIs used to populate these subtrees, see https://github.com/dotnet/project-system/blob/master/docs/repo/dependencies-node-roadmap.md#transitive-dependencies